% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_fpc_scores.R
\name{estimate_fpc_scores}
\alias{estimate_fpc_scores}
\title{estimate_fpc_scores}
\usage{
estimate_fpc_scores(tfb_fpc_obj, Ypred, method = "blup")
}
\arguments{
\item{tfb_fpc_obj}{a fitted FPCA object produced by \code{\link[=rfr_fpca]{rfr_fpca()}}}

\item{Ypred}{Matrix containing curves for which estimated scores are needed}

\item{method}{BLUPs or Numeric integration. Only BLUPs currently implemented}
}
\value{

}
\description{
estimate_fpc_scores
}
