% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fpca-utils.R
\name{extract_fpc_scores}
\alias{extract_fpc_scores}
\title{Extract scores from an FPC object}
\usage{
extract_fpc_scores(rfr_fpca_obj)
}
\arguments{
\item{rfr_fpca_obj}{object returned by \code{rfr_fpca}}
}
\value{
Data frame containing FPCA scores
}
\description{
This function will extract FPC scores from an object produced by \code{rfr_fpca}, and return these scores in a data frame.
}
\examples{
library(refundr)
library(tidyverse)

data(dti_df)
fpca_irregular <- rfr_fpca(Y = "cca", data = dti_df)
scores <- refundr:::extract_fpc_scores(fpca_irregular)

# this gets you scores on a "new" df -- kinda messy though ...
predict(fpca_irregular, slice(dti_df, 1:10)) \%>\%
  refundr:::extract_fpca() \%>\%
  refundr:::extract_fpc_scores()

}
