% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fpca-utils.R
\name{extract_fpca}
\alias{extract_fpca}
\title{Convert tfb_fpc object to a list}
\usage{
extract_fpca(tfb_fpc_obj)
}
\arguments{
\item{tfb_fpc_obj}{object turned by \code{tfb_fpc}}
}
\description{
Convert tfb_fpc object to a list
}
