% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fpca_ssvd.R
\name{fpca_ssvd}
\alias{fpca_ssvd}
\title{Smoothed FPCA via iterative penalized rank one SVDs.}
\usage{
fpca_ssvd(Y = NULL, argvals = NULL, npc = NA, center = TRUE,
  maxiter = 15, tol = 1e-04, diffpen = 3, gridsearch = TRUE,
  alphagrid = 1.5^(-20:40), lower.alpha = 1e-05, upper.alpha = 1e+07,
  verbose = FALSE, integration = "trapezoidal")
}
\arguments{
\item{Y}{data matrix (rows: observations; columns: grid of eval. points)}

\item{argvals}{the argument values of the function evaluations in \code{Y},
defaults to a equidistant grid from 0 to 1. See Details.}

\item{npc}{how many smooth SVs to try to extract, if \code{NA} (the default)
the hard thresholding rule of Donoho, Gavish (2013) is used (see Details,
References).}

\item{center}{center \code{Y} so that its column-means are 0? Defaults to
\code{TRUE}}

\item{maxiter}{how many iterations of the power algorithm to perform at most
(defaults to 15)}

\item{tol}{convergence tolerance for power algorithm (defaults to 1e-4)}

\item{diffpen}{difference penalty order controlling the desired smoothness of
the right singular vectors, defaults to 3 (i.e., deviations from local
quadratic polynomials).}

\item{gridsearch}{use \code{\link[stats:optimize]{stats::optimize()}} or a grid search to find
GCV-optimal smoothing parameters? defaults to \code{TRUE}.}

\item{alphagrid}{grid of smoothing parameter values for grid search}

\item{lower.alpha}{lower limit for for smoothing parameter if
\code{!gridsearch}}

\item{upper.alpha}{upper limit for smoothing parameter if \code{!gridsearch}}

\item{verbose}{generate graphical summary of progress and diagnostic messages?
defaults to \code{FALSE}}

\item{integration}{ignored, see Details.}
}
\description{
Implements the algorithm of Huang, Shen, Buja (2008) for finding smooth right
singular vectors of a matrix \code{X} containing (contaminated) evaluations of
functional random variables on a regular, equidistant grid. If the number of
smooth SVs to extract is not specified, the function hazards a guess for the
appropriate number based on the asymptotically optimal truncation threshold
under the assumption of a low rank matrix contaminated with i.i.d. Gaussian
noise with unknown variance derived in Donoho, Gavish (2013). Please note that
Donoho, Gavish (2013) should be regarded as experimental for functional PCA,
and will typically not work well if you have more observations than grid
points.
}
\references{
Huang, J. Z., Shen, H., and Buja, A. (2008). Functional principal
components analysis via penalized rank one approximation. \emph{Electronic
Journal of Statistics}, 2, 678-695

Donoho, D.L., and Gavish, M. (2013). The Optimal Hard Threshold for Singular
Values is 4/sqrt(3). eprint arXiv:1305.5870. Available from
\url{http://arxiv.org/abs/1305.5870}.
}
\author{
Fabian Scheipl
}
