% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fpca-methods.R
\name{predict.rfr_fpca}
\alias{predict.rfr_fpca}
\title{Prediction for functional principal component analysis}
\usage{
\method{predict}{rfr_fpca}(object, ...)
}
\arguments{
\item{object}{an \code{rf_fpca} object}

\item{...}{optional arguments to be passed to methods}
}
\description{
Takes a fitted \code{rfr_fpca}-object produced by \code{\link{rfr_fpca}()} and
produces predictions given a new set of values or the original values used for the
model fit.
}
