% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refunder-package.R
\docType{package}
\name{refunder}
\alias{refunder}
\alias{refunder-package}
\title{refunder}
\description{
`refunder` makes functional data analyses
in `R` more consistent and user friendly\cr\cr
`refunder` provides:
}
\details{
- tidy interface for common fpca methods and funcional regression
  - plays nice with the `tidyfun` package for wrangling functional data
\cr\cr
\cr
Also see `vignette("refunder")` for a brief introduction.
}
