% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refundr-package.R
\docType{package}
\name{refundr-package}
\alias{refundr}
\alias{refundr-package}
\title{refundr}
\description{
\code{refundr} makes functional data analyses in \code{R} more consistent and
user friendly\cr\cr
\code{refundr} provides:
\itemize{
\item tidy interface for common fpca methods and funcional regression
\item plays nice with the \code{tidyfun} package for wrangling functional data
}

Also see \code{vignette("refundr")} for a brief introduction.
}
\author{
\strong{Maintainer}: Julia Wrobel \email{julia.wrobel@cuanschutz.edu}

Authors:
\itemize{
  \item Fabian Scheipl \email{fabian.scheipl@googlemail.com}
  \item Sara Kim \email{gk2501@cumc.columbia.edu}
  \item Jeff Goldsmith
}

}
\keyword{internal}
