% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refundr-package.R
\docType{package}
\name{refundr}
\alias{refundr}
\title{refundr}
\description{
\code{refundr} makes functional data analyses
in \code{R} more consistent and user friendly\cr\cr
\code{refundr} provides:
\itemize{
\item tidy interface for common fpca methods and funcional regression
\item plays nice with the \code{tidyfun} package for wrangling functional data
}

Also see \code{vignette("refundr")} for a brief introduction.
}
