% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fpca-methods.R
\name{residuals.rfr_fpca}
\alias{residuals.rfr_fpca}
\alias{fitted.rfr_fpca}
\title{Obtain residuals and fitted values for a rfr_fpca models}
\usage{
\method{residuals}{rfr_fpca}(object, ...)

\method{fitted}{rfr_fpca}(object, ...)
}
\arguments{
\item{object}{an \code{rf_fpca} object}

\item{...}{optional arguments to be passed to methods}
}
\description{
Obtain residuals and fitted values for a rfr_fpca models
}
