% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfr_fpca.R
\name{rfr_fpca}
\alias{rfr_fpca}
\title{rfr_fpca Generic function for functional principal component analyses.}
\usage{
rfr_fpca(Y, pve = 0.99, npc = NULL, fpca_method = NULL, ...)
}
\arguments{
\item{Y}{a \code{tfd} data vector.}

\item{pve}{proportion of variance explained: used to choose the number of
principal components.}

\item{npc}{prespecified value for the number of principal components (if
given, this overrides \code{pve}).}

\item{fpca_method}{fpca method of choice. Different options are available but defaults will be set depending on regularity/irregularity of the data}

\item{...}{optional arguments to be passed to methods. Takes arguments from original \code{refund} functions.}
}
\value{

}
\description{
Tidy interface to refund fpca functions
}
\details{
Allows calls to \code{fpca.sc}, \code{fpca.face}, and \code{fpca.ssvd} as implemented in the \code{refund} package. Tidy functional data objects are input and tidy objects are returned as well. Generic function takes in a \code{tfd} vector and passes it to a fpca method.
}
\examples{
\dontrun{
library(refunder)
data(dti_df)
fpca_results <- rfr_fpca(Y = dti_df$cca)

data(chf_df)
fpca_results <- rfr_fpca(Y = chf_df$activity)
}

}
\seealso{
\code{\link{rfr_fpca.tfd_irreg}}, \code{\link{rfr_fpca.tfd_reg}}
}
\author{
Julia, Jeff, Fabian (plus any authors of the original refund code)
}
