% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfr_fpca.tfd_reg.R
\name{rfr_fpca.tfd_reg}
\alias{rfr_fpca.tfd_reg}
\title{Tidy interface for fpca with regularly-spaced functional data}
\usage{
\method{rfr_fpca}{tfd_reg}(Y, pve = 0.99, npc = NULL,
  fpca_method = NULL, ...)
}
\arguments{
\item{Y}{a \code{tfd} data vector.}

\item{pve}{proportion of variance explained: used to choose the number of
principal components.}

\item{npc}{prespecified value for the number of principal components (if
given, this overrides \code{pve}).}

\item{fpca_method}{fpca method of choice. Different options are available but defaults will be set depending on regularity/irregularity of the data}

\item{...}{optional arguments to be passed to methods. Takes arguments from original \code{refund} functions.}
}
\value{

}
\description{
Default call for regular data is to \code{fpca_face}, this function processes the data and provides a wrapper to \code{fpca_face}.
}
\examples{
\dontrun{
library(refunder)
data(chf_df)
fpca_results = rfr_fpca(Y = chf_df$activity)
}
}
\seealso{
\code{\link{rfr_fpca}}, \code{\link{rfr_fpca.tfd_irreg}}
}
\author{
Julia, Jeff, Fabian (plus any authors of the original refund code)
}
