% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_fpca.R
\name{tidy_fpca}
\alias{tidy_fpca}
\title{Tidy interface to refund fpca functions}
\usage{
tidy_fpca(Y, data, pve = 0.99, npc = NULL, method = NULL, ...)
}
\arguments{
\item{Y}{a \code{tfd} data object that exists as a column in the dataframe \code{data} (regular vs. irregular may be assigned differnet defaults).}

\item{data}{a \code{tidyfun} dataframe containing a \code{tfd} vector of data.}

\item{pve}{proportion of variance explained: used to choose the number of
principal components.}

\item{npc}{prespecified value for the number of principal components (if
given, this overrides \code{pve}).}

\item{method}{fpca method of choice. Different options are available but defaults will be set depending on regularity/irregularity of the data}

\item{...}{optional arguments to be passed to methods. Takes arguments from original \code{refund} functions.}
}
\value{

}
\description{
Allows calls to \code{fpca.sc}, \code{fpca.face}, and \code{fpca.ssvd} as implemented in the \code{refund} package. Tidy functional data objects are input and tidy objects are returned as well.
}
\examples{
\dontrun{
library(tidyfun)
library(refunder)
data(dti_df)
fpca_results = tidy_fpca(cca, dti_df)
}

}
\references{
keep original references
}
\author{
Julia, Jeff, Fabian (plus any authors of the original refund code)
}
