% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refunder-package.R
\docType{package}
\name{refunder}
\alias{refunder}
\title{refunder}
\description{
\code{refunder} makes functional data analyses
in \code{R} more consistent and user friendly\cr\cr
\code{refunder} provides:
\itemize{
\item tidy interface for common fpca methods and funcional regression
\item plays nice with the \code{tidyfun} package for wrangling functional data
}

Also see \code{vignette("refunder")} for a brief introduction.
}
