% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfr_fpca.R, R/rfr_fpca.tfd_irreg.R,
%   R/rfr_fpca.tfd_reg.R
\name{rfr_fpca}
\alias{rfr_fpca}
\alias{rfr_fpca.tfb}
\alias{rfr_fpca.tfd_irreg}
\alias{rfr_fpca.tfd_reg}
\title{rfr_fpca Generic function for functional principal component
analyses.}
\usage{
rfr_fpca(Y, data, pve = 0.99, npc = NULL, method = NULL, ...)

\method{rfr_fpca}{tfb}(Y, data, pve = 0.99, npc = NULL,
  method = NULL, ...)

\method{rfr_fpca}{tfd_irreg}(Y, data, pve = 0.99, npc = NULL,
  method = fpca_sc, ...)

\method{rfr_fpca}{tfd_reg}(Y, data, pve = 0.99, npc = NULL,
  method = fpca_face, ...)
}
\arguments{
\item{Y}{character value indicating name for a \code{tfd} data vector. Class \code{tfb} also supported.}

\item{data}{a \code{data.frame} containing the \code{Y} vector as a column.}

\item{pve}{proportion of variance explained: used to choose the number of
principal components.}

\item{npc}{prespecified value for the number of principal components (if
given, this overrides \code{pve}).}

\item{method}{FPCA method of choice. Different options are available but
default depends on regularity/irregularity of \code{data}, see \strong{Methods}.}

\item{...}{optional arguments to be passed to FPCA algorithms. Takes arguments from
original \code{refund} functions.}
}
\value{
an object of class \code{fpca}, internal structure not yet fixed. Currently
a list with a \code{\link[tidyfun:tfb_fpc]{tidyfun::tfb_fpc()}} object and additional entries describingt the FPCA results.
}
\description{
Tidy interface to \code{refund} FPCA functions
}
\details{
Allows calls to (wrappers for) \code{\link[refund:fpca.sc]{refund::fpca.sc()}}, \code{\link[refund:fpca.face]{refund::fpca.face()}}, and
\code{\link[refund:fpca.ssvd]{refund::fpca.ssvd()}} as implemented in the \code{{refund}} package. Tidy functional data
objects are input and tidy objects are returned as well.
}
\section{Methods (by class)}{
\itemize{
\item \code{tfd_irreg}: FPCA for data on irregular grids defaults to \code{\link[refund:fpca.sc]{refund::fpca.sc()}}

\item \code{tfd_reg}: FPCA for data on a regular grid defaults to \code{\link[refund:fpca.face]{refund::fpca.face()}}
}}

\examples{
\dontrun{
library(refunder)
data(dti_df)
fpca_irregular <- rfr_fpca(Y = "cca", data = dti_df)

data(chf_df)
fpca_regular <- rfr_fpca(Y = "activity", data = chf_df)
}

}
\author{
Julia, Jeff, Fabian (plus any authors of the original refund code)
}
